- Add utils.py     - Changed all classes using requests to use utils#request
